requirements.txt
gradio==3.45.2
pydub==0.25.1
scipy==1.11.3
streamlit==1.27.0
streamlit_audiorecorder==0.0.3.post1
torch==2.0.1
openai-whisper